continuing testing and correcting imports

+++ src/modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
-import { deleteFile } from '../../../../utils/file';
-import { IUsersRepository } from '../../repositories/IUsersRepository';
+import { IUsersRepository } from '@modules/accounts/repositories/IUsersRepository';
+import { deleteFile } from '@utils/file';
 
 interface IRequest {
     user_id: string;
@@ -18,13 +18,15 @@ class UpdateUserAvatarUseCase {
     async execute({ user_id, avatar_file }: IRequest): Promise<void> {
         const user = await this.usersRepository.findById(user_id);
 
-        if (user.avatar) {
-            await deleteFile(`./tmp/avatar/${user.avatar}`);
+        if (user) {
+            if (user.avatar) {
+                await deleteFile(`./tmp/avatar/${user.avatar}`);
+            }
+
+            user.avatar = avatar_file;
+
+            await this.usersRepository.create(user);
         }
-
-        user.avatar = avatar_file;
-
-        await this.usersRepository.create(user);
     }
 }
